+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.1.0] - 2021-02-27
+
 ### Added
 
 - `setAlignment()` function
@@ -121,7 +123,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Lyrics synchronizer
 - Three sizes: tiny, medium, full.
 
-[unreleased]: https://github.com/guoyunhe/rabbit-lyrics/compare/v2.0.0...HEAD
+[unreleased]: https://github.com/guoyunhe/rabbit-lyrics/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/guoyunhe/rabbit-lyrics/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/guoyunhe/rabbit-lyrics/compare/v1.3.0...v2.0.0
 [1.3.0]: https://github.com/guoyunhe/rabbit-lyrics/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/guoyunhe/rabbit-lyrics/compare/v1.1.2...v1.2.0